refactor: giveResponse function refactor: getRequest
