+++ src/getUser.ts
@@ -2,13 +2,9 @@ import http from 'http';
 import users from './index';
 import idValidate from './uidvalidate';
 import * as err from './errors';
+import giveResponse from './giveresponse';
 
-const giveResponse = async (res: http.ServerResponse, data: any) => {
-    res.writeHead(200, { 'Content-Type': 'application/json' });
-    res.end(JSON.stringify(data));
-};
-
-const getUser = async (req: http.IncomingMessage, res: http.ServerResponse) => {
+const getUser = async (req: http.IncomingMessage, res: http.ServerResponse): Promise<void> => {
     if (req.url === '/api/users') {
        giveResponse(res, users);
     } else if (req.url!.match(/\/api\/users\/\w+/)) {
